setup/requirements.txt
pymongo
pandas
python-dotenv
google-genai